chef reports error when '$ vagrant halt' and then '$ vagrant up': New better way of ssh config [#23]
5. **How to use ssh keys on the virtual machine**?

``teracy-dev/home/.ssh`` on the host machine and ``~/.ssh`` on the virtual machine are in sync. You
could copy your existing ssh keys into one location and it will be available in the other location.

For example, to use existing ssh keys, type the following commands on the host machine terminal
window:
::
    $ cd teracy-dev
    $ cp ~/.ssh/id_rsa* home/.ssh

Or to create new ssh keys on the virtual machine, just create it and these keys will be copied into
``teracy-dev/home/.ssh``.